# Optimize BaseParser#unnormalize method to replace "\r\n" with "\n" only when "\r\n" is included